Tidy

+++ src/de/lmu/ifi/dbs/elki/visualization/visualizers/visunproj/PixmapVisualizer.java
@@ -68,7 +68,7 @@ public class PixmapVisualizer extends AbstractVisFactory {
     for(PixmapResult pr : prs) {
       // Add plots, attach visualizer
       final VisualizationTask task = new VisualizationTask(NAME, pr, null, this);
-      task.width = pr.getImage().getWidth() / pr.getImage().getHeight();
+      task.width = pr.getImage().getWidth() / (double) pr.getImage().getHeight();
       task.height = 1.0;
       task.put(VisualizationTask.META_LEVEL, VisualizationTask.LEVEL_STATIC);
       baseResult.getHierarchy().add(pr, task);
